requirements.txt
pandas
scikit-learn
joblib
scipy
crunch-cli
lightgbm
xgboost